src/requirements.txt
discord
ffmpeg-python
pyyaml
pynacl
python-dotenv
